+++ CheeseQuest2/Main.java
@@ -1,16 +1,7 @@
-/**
- * Start game here
- */
-
-import menu.*;
+import gui.MainFrame;
 
 public class Main {
-    public static Menu menu;
     public static void main(String[] args) {
-        menu = new MainMenu();
-        while (menu != null) {
-            menu.start();
-            menu = menu.getNextMenu();
-        }
-  }
+        MainFrame frame = new MainFrame();
+    }
 }