+++ CheeseQuest2/game/Game.java
@@ -1,5 +1,8 @@
 package game;
 
+import game.system.Player;
+import game.system.World;
+
 /**
  * Manages InputPanel input to determine player and world interactions
  */
@@ -7,7 +10,7 @@ public class Game {
     private Player player;
     private World world;
 
-    
+
     /**
      * Input string determines game actions
      * @param String input from InputPanel
@@ -16,5 +19,17 @@ public class Game {
         paritionInput();
         processInput();
         sendOutput();
+    }
+
+    private void paritionInput() {
+
+    }
+
+    private void processInput() {
+
+    }
+
+    private void sendOutput() {
+        
     }
 }