+++ CheeseQuest2/game/menu/LoadMenu.java
@@ -4,6 +4,8 @@ import java.io.*;
 import java.util.*;
 import file.*;
 
+import game.system.*;
+
 /**
  * Prompts to load game save
  * Uses ReadObject to detect and load games
@@ -17,6 +19,9 @@ public class LoadMenu extends Menu {
     private ReadObject reader;
 
 
+    /**
+     * Default LoadMenu constructor
+     */
     private LoadMenu() {
         existingSave = new ArrayList<String>();
         validSave = new ArrayList<String>();
@@ -27,6 +32,20 @@ public class LoadMenu extends Menu {
 
     public static LoadMenu getInstance() {
         return loadMenu;
+    }
+
+    @Override
+    public void outputPrompt() {
+        outputln("Load Game");
+        String gameName;
+        for (int i = 0; i < validSave.size(); i++) {
+            if (existingSave.contains(validSave.get(i))) {
+                gameName = "Load game"; // Game exists
+            } else {
+                gameName = "New game"; // Game does not exist
+            }
+            outputln("    " + i + ". " + gameName);
+        }
     }
 
     @Override
@@ -47,16 +66,21 @@ public class LoadMenu extends Menu {
     public void determineExistingSaves() {
         existingSave.clear();
         for (int i = 1; i <= MAXIMUM_SAVES; i++) {
-            setSaveNum(i);
+            reader.setSaveNum(i);
             if (reader.saveExists()) {
                 existingSave.add(Integer.toString(i));
             }
         }
     }
 
+    /**
+     * Sets the GameMenu's world to saveNum
+     * @param int saveNum of world to load
+     */
     public void loadGame(int saveNum) {
         reader.setSaveNum(saveNum);
         World world = (World) reader.deserialize();
-        GameMenu.setWorld(world);
+        GameMenu gameMenu = GameMenu.getInstance();
+        gameMenu.setWorld(world);
     }
 }