+++ CheeseQuest2/game/object/room/Hallway.java
@@ -1,10 +1,15 @@
 package game.object.room;
 
 public class Hallway extends Room {
+	private static Hallway hallway = new Hallway();
 	private Hallway() {
 		addSingleName(new String[] {"hallway"});
 		addDescription(new String[] {"It's long"});
-		
-		coordinates.setWestRoom(JailCell.getInstance());
+
+		getCoordinates().setWestRoom(JailCell.getInstance());
+	}
+
+	public static Hallway getInstance() {
+		return hallway;
 	}
 }