+++ CheeseQuest2/game/object/room/JailCell.java
@@ -1,10 +1,15 @@
 package game.object.room;
 
 public class JailCell extends Room {
+	private static JailCell jailCell = new JailCell();
 	private JailCell() {
 		addSingleName(new String[] {"jail cell","cell"});
 		addDescription(new String[] {"It's cold and dusty."});
-		
-		coordinates.setEastRoom(Hallway.getInstance());
+
+		getCoordinates().setEastRoom(Hallway.getInstance());
+	}
+
+	public static JailCell getInstance() {
+		return jailCell;
 	}
 }